# Conflicts:
#	src/main/java/seedu/address/model/person/Person.java

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -10,7 +10,8 @@ import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
-import seedu.address.ui.DeletePopup;
+import seedu.address.ui.DeletePopupHandler;
+import seedu.address.ui.PopupHandler;
 
 /**
  * Deletes a person identified using it's displayed index from the address book.
@@ -29,23 +30,29 @@ public class DeleteCommand extends Command {
     private final Index targetIndex;
     private final String targetName;
     private final boolean isDeletedByName;
+    private final PopupHandler infoPopupHandler;
+    private final DeletePopupHandler deletePopupHandler;
 
     /**
      * Creates a DeleteCommand to delete by index.
      */
-    public DeleteCommand(Index targetIndex) {
+    public DeleteCommand(Index targetIndex, PopupHandler infoPopupHandler, DeletePopupHandler deletePopupHandler) {
         this.targetIndex = targetIndex;
         this.targetName = null;
         this.isDeletedByName = false;
+        this.infoPopupHandler = infoPopupHandler;
+        this.deletePopupHandler = deletePopupHandler;
     }
 
     /**
      * Creates a DeleteCommand to delete by name.
      */
-    public DeleteCommand(String targetName) {
+    public DeleteCommand(String targetName, PopupHandler infoPopupHandler, DeletePopupHandler deletePopupHandler) {
         this.targetIndex = null;
         this.targetName = targetName;
         this.isDeletedByName = true;
+        this.infoPopupHandler = infoPopupHandler;
+        this.deletePopupHandler = deletePopupHandler;
     }
 
     @Override
@@ -55,12 +62,16 @@ public class DeleteCommand extends Command {
 
         // delete by name
         if (isDeletedByName) {
+            assert targetName != null : "Target name should not be null when deleting by name";
             List<Person> exactMatches = lastShownList.stream()
                     .filter(p -> p.getName().fullName.equalsIgnoreCase(targetName))
                     .toList();
 
             if (exactMatches.size() == 1) {
                 Person personToDelete = exactMatches.get(0);
+                if (isDeletionCancelled(personToDelete)) {
+                    throw new CommandException("Deletion cancelled.");
+                }
                 model.deletePerson(personToDelete);
                 return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS,
                         Messages.format(personToDelete)));
@@ -74,11 +85,14 @@ public class DeleteCommand extends Command {
                     : exactMatches;
 
             if (possibleMatches.isEmpty()) {
-                showNoMatchPopup();
+                infoPopupHandler.showMessage("No matches found. Please try again.");
                 throw new CommandException("Deletion cancelled.");
             }
 
             Person selectedPerson = showDeletePopup(possibleMatches);
+            if (isDeletionCancelled(selectedPerson)) {
+                throw new CommandException("Deletion cancelled.");
+            }
             model.deletePerson(selectedPerson);
             return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS,
                     Messages.format(selectedPerson)));
@@ -86,10 +100,15 @@ public class DeleteCommand extends Command {
 
         // delete by index
         if (targetIndex.getZeroBased() >= lastShownList.size()) {
+            infoPopupHandler.showMessage("Invalid index. Please try again.");
             throw new CommandException(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
+        assert personToDelete != null : "Person to delete should not be null";
+        if (isDeletionCancelled(personToDelete)) {
+            throw new CommandException("Deletion cancelled.");
+        }
         model.deletePerson(personToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
     }
@@ -98,23 +117,15 @@ public class DeleteCommand extends Command {
      * Shows a Delete Pop up for the user to select from matchingResults.
      * */
     private Person showDeletePopup(List<Person> matchingResults) throws CommandException {
-        DeletePopup deletePopup = new DeletePopup();
-        deletePopup.show("Multiple matches found. Type index and ENTER to delete or ESC to cancel:",
-                matchingResults);
-
-        if (deletePopup.isConfirmed()) {
-            return deletePopup.getSelectedPerson();
-        } else {
-            throw new CommandException("Deletion cancelled.");
-        }
+        return deletePopupHandler.showDeletePopup(
+                "Possible matches found below.\nType INDEX and ENTER to delete or ESC to cancel:", matchingResults);
     }
 
     /**
-     * Shows a Delete Pop up to tell the user no match found.
+     * Shows a Confirm Pop up to ask the user whether to proceed with the deletion.
      */
-    private void showNoMatchPopup() {
-        DeletePopup deletePopup = new DeletePopup();
-        deletePopup.show("No matches found. Press ESC to exit.", List.of());
+    private boolean isDeletionCancelled(Person person) {
+        return !deletePopupHandler.confirmDeletion(person);
     }
 
     @Override
